Disagreement with myself over minimum SCLK wait
In the CPHA=1 case, the clock must wait after SS transitions until it's first
transition.  Previously I had this set to the smallest quanta of time I could
(a single fpga_clock cycle), but upon further thought I think it should be held
for a full SPI clock half-period.

The reasoning is that if there is a single fpga_clock between transactions, but
the SPI clock frequency is much lower, then there is a case where the SCLK line
transitions faster than it should (faster than it's frequency).

Now in this case the SPI clock will have a half period that's longer by a single
fpga_clock cycle, so at its fastest it is at the configured SPI clock speed.